pyproject.toml
[tool.poetry]
name = "bookingsystem"
version = "0.1.0"
description = ""
authors = ["nurai5 <nursultanaibekuulu1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pymongo = "^4.6.2"
python-dotenv = "^1.0.1"
mongoengine = "^0.28.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.6.3"}
argon2-cffi = "^23.1.0"
motor = "^3.3.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
beanie = "^1.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
